+++ packages/fluentui/react-northstar/src/components/Toolbar/ToolbarMenuDivider.tsx
@@ -1,5 +1,5 @@
 import { Accessibility } from '@fluentui/accessibility';
-import { getElementType, getUnhandledProps, useAccessibility, useStyles, useTelemetry } from '@fluentui/react-bindings';
+import { getElementType, useUnhandledProps, useAccessibility, useStyles, useTelemetry } from '@fluentui/react-bindings';
 import { mergeComponentVariables } from '@fluentui/styles';
 import * as React from 'react';
 // @ts-ignore
@@ -49,7 +49,7 @@ const ToolbarMenuDivider: React.FC<WithAsProp<ToolbarMenuDividerProps>> &
   });
 
   const ElementType = getElementType(props);
-  const unhandledProps = getUnhandledProps(ToolbarMenuDivider.handledProps, props);
+  const unhandledProps = useUnhandledProps(ToolbarMenuDivider.handledProps, props);
 
   const element = <ElementType {...getA11yProps('root', { ...unhandledProps, className: classes.root })} />;
   setEnd();