+++ packages/fluentui/react-northstar/src/components/Tree/TreeItem.tsx
@@ -1,5 +1,5 @@
 import { Accessibility, treeItemBehavior, TreeItemBehaviorProps } from '@fluentui/accessibility';
-import { getElementType, getUnhandledProps, useAccessibility, useStyles, useTelemetry } from '@fluentui/react-bindings';
+import { getElementType, useUnhandledProps, useAccessibility, useStyles, useTelemetry } from '@fluentui/react-bindings';
 import * as customPropTypes from '@fluentui/react-proptypes';
 import * as _ from 'lodash';
 import * as PropTypes from 'prop-types';
@@ -197,7 +197,7 @@ const TreeItem: React.FC<WithAsProp<TreeItemProps>> &
   });
 
   const ElementType = getElementType(props);
-  const unhandledProps = getUnhandledProps(TreeItem.handledProps, props);
+  const unhandledProps = useUnhandledProps(TreeItem.handledProps, props);
 
   const element = (
     <ElementType