+++ packages/office-ui-fabric-react/src/components/Dropdown/Dropdown.doc.tsx
@@ -19,7 +19,7 @@ export const DropdownPageProps: IDocPageProps = {
   title: 'Dropdown',
   componentName: 'Dropdown',
   componentUrl:
-    'https://github.com/OfficeDev/office-ui-fabric-react/tree/master/packages/office-ui-fabric-react/src/components/Dropdown',
+    'https://github.com/microsoft/fluentui/tree/master/packages/office-ui-fabric-react/src/components/Dropdown',
   examples: [
     {
       title: 'Basic Dropdowns',