+++ packages/office-ui-fabric-react/src/components/MarqueeSelection/MarqueeSelection.doc.tsx
@@ -9,7 +9,7 @@ export const MarqueeSelectionPageProps: IDocPageProps = {
   title: 'MarqueeSelection',
   componentName: 'MarqueeSelection',
   componentUrl:
-    'https://github.com/OfficeDev/office-ui-fabric-react/tree/master/packages/office-ui-fabric-react/src/components/MarqueeSelection',
+    'https://github.com/microsoft/fluentui/tree/master/packages/office-ui-fabric-react/src/components/MarqueeSelection',
   examples: [
     {
       title: 'Basic Selection Example',